refactor: removed updated message renderer provider
