+++ src/components/Core/TextChannelContainer/MessageContainer.tsx
@@ -6,7 +6,6 @@ import { useAppRouter } from '@hooks/useAppRouter';
 import { useMessages } from '@hooks/useMessages';
 import { useMessageSubscription } from '@hooks/useMessageSubscription';
 import { StateMessages } from 'types/messages.types';
-import { MessageRenderer } from '@components/Providers/MessageRenderer';
 import * as Styles from './styles';
 
 interface MessageContainerProps {
@@ -43,25 +42,23 @@ export const MessageContainer = ({ channelIsThread }: MessageContainerProps) => 
   }, [isListRendered]);
 
   return (
-    <MessageRenderer>
-      <Styles.MessageWrapper
-        draggable={false}
-        membersListOpen={isMembersListOpen}
-        mobile={{
-          '@initial': false,
-          '@small': true
-        }}
-      >
-        <MessagesListRenderer
-          startReached={loadMoreMessages}
-          messages={groupedMessages}
-          isReady={isReady}
-          firstItemIndex={firstItemIndex}
-          handleBottomStateChanged={handleBottomStateChanged}
-        />
+    <Styles.MessageWrapper
+      draggable={false}
+      membersListOpen={isMembersListOpen}
+      mobile={{
+        '@initial': false,
+        '@small': true
+      }}
+    >
+      <MessagesListRenderer
+        startReached={loadMoreMessages}
+        messages={groupedMessages}
+        isReady={isReady}
+        firstItemIndex={firstItemIndex}
+        handleBottomStateChanged={handleBottomStateChanged}
+      />
 
-        <TextBox />
-      </Styles.MessageWrapper>
-    </MessageRenderer>
+      <TextBox />
+    </Styles.MessageWrapper>
   );
 };
