+++ src/hooks/useMessages.ts
@@ -7,7 +7,6 @@ import { APIMessage } from 'discord-api-types/v10';
 import { convertMessageToDiscord } from '@util/convertToDiscord/convertMessageToDiscord';
 import { messagesQuery } from '@hooks/messagesQuery';
 import { StateMessages } from 'types/messages.types';
-import { useStoreActions } from '@state';
 
 type MessageState = {
   groupedMessages: APIMessage[][];
@@ -36,7 +35,6 @@ export const useMessages = ({
   });
 
   const [newMessageGroupLength, setNewMessageGroupLength] = useState(0);
-  const addMember = useStoreActions(state => state.guild.addMember);
 
   const [{ data }, fetchHook] = useQuery({
     query: messagesQuery,
@@ -44,18 +42,6 @@ export const useMessages = ({
   });
 
   const isReady = data?.channelV2.id === channel;
-
-  const getMentionedMembers = (msgs: Message[]) => {
-    for (const msg of msgs) {
-      const { mentions } = msg;
-
-      if (!mentions.length) continue;
-
-      for (const member of mentions) {
-        addMember(member);
-      }
-    }
-  };
 
   useEffect(() => {
     if (variables.channel !== channel || variables.threadId !== threadId) {
@@ -75,7 +61,6 @@ export const useMessages = ({
       if (isReadyWithMessages) {
         setNewMessageGroupLength(groupMessages(msgs).length);
 
-        getMentionedMembers(msgs);
         setMessages(prev => [...msgs, ...prev]);
       }
     }
