+++ src/pages/channels/[guild]/[channel].tsx
@@ -7,6 +7,9 @@ import * as Styles from '@components/Core/styles';
 import { useContextMenu } from '@hooks/useContextMenu';
 import dynamic from 'next/dynamic';
 import { GuestFormModal } from '@components/Overlays/Modal/GuestFormModal';
+import { styled } from '@stitches';
+import { MessageRendererProvider } from '@widgetbot/message-renderer';
+import { svgUrls } from '@svg-assets';
 
 // dynamic imports since they are conditionally rendered, helps with bundle size
 const ChannelTopicModal = dynamic(() =>
@@ -23,6 +26,14 @@ const ContextMenu = dynamic(() =>
   import('@components/Overlays/ContextMenu').then(mod => mod.ContextMenu)
 );
 
+const MessageRendererRoot = styled('div', {
+  height: '100%',
+  width: '100%',
+  '*': {
+    fontFamily: 'GgSans'
+  }
+});
+
 function GuildChannel() {
   const { disableBrowserMenu } = useContextMenu();
 
@@ -30,20 +41,48 @@ function GuildChannel() {
 
   const showContextMenu = useStoreState(state => state.ui.showContextMenu);
   const showTopicModal = useStoreState(state => state.ui.showTopicModal);
+  const guildChannels = useStoreState(state => state.guild.guildChannels);
 
   return (
-    <Styles.Main onContextMenu={disableBrowserMenu}>
-      {showContextMenu && <ContextMenu />}
+    <MessageRendererProvider
+      messageButtons={() => []}
+      currentUser={() => null}
+      // @ts-ignore
+      resolveChannel={id => (guildChannels[id] as APIChannel) ?? null}
+      resolveGuild={() => null}
+      resolveMember={() => null}
+      resolveRole={() => null}
+      resolveUser={() => null}
+      svgUrls={svgUrls}
+      seeThreadOnClick={(messageId, thread) =>
+        alert(`See Thread "${thread.name}" clicked on message ${messageId}`)
+      }
+      userMentionOnClick={user =>
+        alert(`User "${user?.global_name ?? user?.username}" mention clicked!`)
+      }
+      roleMentionOnClick={role => alert(`Role "${role.name}" mention clicked!`)}
+      channelMentionOnClick={channel => alert(`Channel "${channel.name}" mention clicked!`)}
+      messageComponentButtonOnClick={(message, customId) => {
+        alert(`Button by custom id "${customId}" pressed on message ${message.id}!`);
+      }}
+    >
+      {({ themeClass }) => (
+        <MessageRendererRoot className={themeClass}>
+          <Styles.Main onContextMenu={disableBrowserMenu}>
+            {showContextMenu && <ContextMenu />}
 
-      <Styles.InnerMain>
-        {showTopicModal && <ChannelTopicModal />}
-        <GuestFormModal />
-        <ChannelsSidebar />
+            <Styles.InnerMain>
+              {showTopicModal && <ChannelTopicModal />}
+              <GuestFormModal />
+              <ChannelsSidebar />
 
-        <TextChannelContainer />
-      </Styles.InnerMain>
-      {isDomThreadsPanelOpen && <ThreadPanel />}
-    </Styles.Main>
+              <TextChannelContainer />
+            </Styles.InnerMain>
+            {isDomThreadsPanelOpen && <ThreadPanel />}
+          </Styles.Main>
+        </MessageRendererRoot>
+      )}
+    </MessageRendererProvider>
   );
 }
